Shooting pt2.
I have been working on the shooting. Since my last update I have been able to get the bullet to spawn. However, the bullet just seems to fall and not go that far.

Note: if any of you want to change anything in the scripts; DO NOT RENAME THE SCRIPT ITSELF.
       ~Joey.